server/requirements.txt
requests 
Django==4.1.3
djangorestframework==3.14.0
django-cors-headers==3.13.0
beautifulsoup4==4.10.0
requests-html
psycopg2
dj-database-url
gunicorn==20.1.0
djangorestframework-simplejwt
Pillow==9.4.0
pytest==7.2.2